Move UpdateBLSConfigurationTask to payload base
This task is used by live_image and live_os payloads so move it to the
base.
